requirements.txt

# CR18 Visual Workflow Designer - Python Dependencies
# Core Flask/Web framework
flask==3.0.0
flask-cors==4.0.0

# AI/LLM Integration
anthropic==0.40.0

# Environment management
python-dotenv==1.0.0

# Data processing and utilities
lxml==4.9.3
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv
